feat(workflow): add Ready-for-Merge human gate (#935)
Wire the dead ready_to_merge state to an actual transition edge by
introducing HumanMergeApproved event kind and the record_merge_approved
store method.  Add POST /tasks/{id}/merge endpoint so the dashboard
Merge button can promote a workflow from ReadyToMerge → Done.  Add
require_human_gate_before_merge config flag (default false) to
IssueWorkflowPolicy for future enforcement.

- harness-workflow: split issue_lifecycle.rs (1037 lines) into typed
  issue_lifecycle.rs + IssueWorkflowStore in issue_workflow_store.rs +
  tests in issue_workflow_store_tests.rs to stay under U-16 800-line cap
- HumanMergeApproved apply_event arm guards on ReadyToMerge; warns and
  short-circuits on any other state (no last_event mutation)
- harness-core: require_human_gate_before_merge: bool with serde default
- harness-server: task_mutation_routes.rs (new) hosts merge_task handler;
  validates task pr_url, workflow state, then calls record_merge_approved
- frontend Active.tsx: Merge button in Ready column, POST /tasks/{id}/merge,
  invalidates tasks query on success

Closes #935
