+++ crates/harness-server/src/http/http_router.rs
@@ -10,7 +10,7 @@ use super::{
     auth, get_issue_workflow_by_issue, get_issue_workflow_by_pr, get_project_workflow_by_project,
     get_task, get_task_artifacts, get_task_prompts, github_webhook, handle_rpc, health_check,
     ingest_signal, intake_status, list_tasks, password_reset, project_queue_stats, state::AppState,
-    stream_task_sse, task_routes,
+    stream_task_sse, task_mutation_routes, task_routes,
 };
 
 pub(super) fn build_router(state: Arc<AppState>) -> Router {
@@ -31,6 +31,7 @@ pub(super) fn build_router(state: Arc<AppState>) -> Router {
         .route("/tasks/batch", post(task_routes::create_tasks_batch))
         .route("/tasks/{id}", get(get_task))
         .route("/tasks/{id}/cancel", post(task_routes::cancel_task))
+        .route("/tasks/{id}/merge", post(task_mutation_routes::merge_task))
         .route("/tasks/{id}/artifacts", get(get_task_artifacts))
         .route("/tasks/{id}/prompts", get(get_task_prompts))
         .route("/tasks/{id}/stream", get(stream_task_sse))
