+++ crates/harness-workflow/src/lib.rs
@@ -6,6 +6,7 @@
 pub mod checkpoint;
 pub mod circuit_breaker;
 pub mod issue_lifecycle;
+pub mod issue_workflow_store;
 pub mod plan_db;
 pub mod project_lifecycle;
 pub mod task_queue;
